Move receive_imf tests into a separate file
